+++ src/main/java/cinnamon/world/entity/vehicle/Vehicle.java
@@ -5,6 +5,7 @@ import cinnamon.utils.Rotation;
 import cinnamon.world.entity.Entity;
 import cinnamon.world.entity.PhysEntity;
 import cinnamon.world.entity.living.LivingEntity;
+import org.joml.Math;
 import org.joml.Vector3f;
 
 import java.util.UUID;
@@ -35,7 +36,7 @@ public abstract class Vehicle extends PhysEntity {
         impulse.mul(getMoveSpeed());
 
         //move the entity in facing direction
-        this.impulse.rotateY((float) Math.toRadians(-rot.y));
+        this.impulse.rotateY(Math.toRadians(-rot.y));
     }
 
     @Override
